EXT-AUDIT-001: Memory Cleansing is Fragile and Partly Ineffective
This is a general problem with things written in Go.  While the audit
recommended that it be removed mostly due to being a lost cause, my
current thoughts are that some sanitization is better than no
sanitization.

This may be revisited at a later date, but for now add a note in the
documentation and comments where appropriate.