extra/x25519: Add Ed25519->X25519 key conversion helpers
Rescue the Ed25519->X25519 key conversion helpers from agl's deprecated
repository.  While some would argue that this method of doing things is
unnecessary and or unsafe, they are arguably useful in certain
contexts.

Note: This is not quite a drop-in replacement for code that used agl's
old routines as it is assumed that the keys will be fed into the
standard library's new X25519 function.