ed25519: Reject small order A
libsodium does this, we do it in oasis-core, it is a good idea (needed
for non-repudiation) in general.

This diverges from upstream behavior, but that bridge was burnt the
moment that we switched to cofactored verification anyway, and no one
should be using small order public keys in the first place.